Start splitting up the AtomSpace into two parts
